refactoring CaeSolver is done, femQuickAnalys is fine
